pyproject.toml
[tool.poetry]
name = "gcs-uploader"
version = "0.0.1"
description = "Upload files to Google Cloud Storage"
authors = ["Daniel Komesu <danielkomesu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-storage = "^2.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
